pyproject.toml
[project]
name = "imapfs"
version = "0.0.1"
description = "Add your description here"
readme = "README.md"
authors = [{ name = "Reuben Frankel", email = "rfrankel@matatika.com" }]
requires-python = ">=3.9"
dependencies = [
    "fsspec>=2025.5.1",
    "imap-tools>=1.11.0",
]

[dependency-groups]
test = [
    "pytest>=8",
    "pytest-github-actions-annotate-failures>=0.3",
    "python-dotenv>=1.1.1",
]

[tool.ruff]
src = ["impfs"]
target-version = "py39"

[tool.ruff.lint]
ignore = [
    "ANN001",  # missing-type-function-argument
    "ANN002",  # missing-type-args
    "ANN003",  # missing-type-kwargs
    "ANN201",  # missing-return-type-undocumented-public-function
    "ANN202",  # missing-return-type-private-function
    "COM812",  # missing-trailing-comma
]
select = ["ALL"]

[tool.ruff.lint.flake8-annotations]
allow-star-arg-any = true

[tool.ruff.lint.pydocstyle]
convention = "google"
